Implement CBOR encoder (currently subset)
For simplicity we don't implement everything right away to be able to
demonstrate the RPC proof of concept.

In particular, it would be nice to have an iterator for collections
(lists/sets). We also don't implement tags, but the possibility is there
for future expansion if needed.

And yes the test/cbor.cpp parses the table in comms-ccf/cbor.hpp from
appendix A, which was a bit of a pain to write.