FreeRTOS-Demo: Add RPC example
This adds comms-ccf onto the example. First, the link script changes
are required for the static initializer of Rpc/Ccf.

We also need the `-fno-use-cxa-atexit` for similar reasons.

There is some handle-turning for CMakePresets.txt to make it run different
build types as a workflow.

Then we remove the hello message and add UART0 as the comms-ccf transport.

Then comms-ccf.cpp contains the main part of using it.

Hopefully it is mostly straight forward.