Add simple printf logging
This does pull printf in, deferred formatting (and using pointers for
flash data) could be used to make it work better.

For the deferred formatting, we would need to send structured data, but
CBOR and UTF-8 can co-exist as 0x8X/0x9X is an invalid UTF-8 sequence
so the two can always be told apart.