Remove nosys linker warnings by crashing
The linker is complaining about these functions because in the nosys
library they are implemented to always fail. This is fine, we don't need
the warnings, in fact we implement them as not functions that just fail
but instead crash if they are used. This way we know they aren't used
(in the code we run at least).