Add mapfile parser to aid comparing binary sizes
The problem with the current format is it includes the memory address
which changes too much between builds.

This adds a tool to print YAML which skips memory addresses so is
easily amenable to `diff`, and an SVG square map format that might help
visualisation.

The SVG generation is not the neatest code, lots of ad-hoc things to
try and make the output a bit better based on use, but it's okay and
the output is not too bad.