Measure size costs in more detail
This adds functionality to compare mapfiles to the mapfile parser. The
code is again more of a quick script than pretty, but it works.

This makes it easier to reproduce and compare than using different
git revisions.

It adds the computed SVGs/markdown file even though it is a build output,
as I don't have CI yet. And it's under 3MiB anyway.