Auto-generate comparison table
This also uses just the section/area names for the IDs, which assumes
they won't clash, not a guarantee but very likely for embedded because
of `-ffunction-sections` and `-fdata-sections`, and if they do then all
that happens is the linking will be off. Unfortunate but not the end of
the world.